+++ e2e/src/app.e2e-spec.ts
@@ -4,13 +4,8 @@ import { browser, logging } from 'protractor';
 describe('workspace-project App', () => {
   let page: AppPage;
 
-  beforeEach(() => {
+  beforeEach((): any => {
     page = new AppPage();
-  });
-
-  it('should display welcome message', () => {
-    page.navigateTo();
-    expect(page.getTitleText()).toEqual('Welcome to weatherApp!');
   });
 
   afterEach(async () => {